refactor: move prove_and_verify_vm_minimal to test_utils
Remove duplication between segment_tests.rs and prove_elfs_tests.rs
by moving the shared helper to test_utils. This also removes the need
for type aliases in segment_tests.rs.
