+++ prover/src/tests/prove_elfs_tests.rs
@@ -19,24 +19,17 @@ use stark::constraints::transition::TransitionConstraint;
 use stark::lookup::{AirWithBuses, AuxiliaryTraceBuildData};
 use stark::proof::options::ProofOptions;
 use stark::prover::{IsStarkProver, Prover};
-use stark::trace::TraceTable;
 use stark::traits::AIR;
 use stark::verifier::{IsStarkVerifier, Verifier};
 
 use crate::tables::bitwise::{generate_bitwise_trace, update_multiplicities};
 use crate::tables::lt::generate_lt_trace;
 use crate::tables::trace_builder::Traces;
-use crate::tables::types::{GoldilocksExtension, GoldilocksField};
-
-// Import shared utilities
 use crate::test_utils::{
     collect_bitwise_lookups_from_logs, collect_bitwise_lookups_from_lt,
     collect_lt_lookups_from_logs, create_bitwise_air, create_cpu_air, create_lt_air,
-    generate_minimal_bitwise_trace, run_asm_elf,
+    generate_minimal_bitwise_trace, prove_and_verify_vm_minimal, run_asm_elf, E, F,
 };
-
-type F = GoldilocksField;
-type E = GoldilocksExtension;
 
 /// Alias for compatibility with existing test code.
 fn collect_bitwise_lookups(
@@ -48,8 +41,6 @@ fn collect_bitwise_lookups(
     collect_bitwise_lookups_from_logs(logs, instructions)
 }
 
-// AIR creation helpers and lookup collection functions are now in test_utils module
-
 // =============================================================================
 // Prover test helpers
 // =============================================================================
@@ -59,9 +50,9 @@ fn collect_bitwise_lookups(
 /// Uses the FULL 2^20 row bitwise table with preprocessed commitment.
 /// Returns true if verification succeeds.
 fn prove_and_verify_vm(
-    cpu_trace: &mut TraceTable<F, E>,
-    bitwise_trace: &mut TraceTable<F, E>,
-    lt_trace: &mut TraceTable<F, E>,
+    cpu_trace: &mut stark::trace::TraceTable<F, E>,
+    bitwise_trace: &mut stark::trace::TraceTable<F, E>,
+    lt_trace: &mut stark::trace::TraceTable<F, E>,
 ) -> bool {
     let proof_options = ProofOptions::default_test_options();
 
@@ -69,15 +60,12 @@ fn prove_and_verify_vm(
     let bitwise_air = create_bitwise_air(&proof_options);
     let lt_air = create_lt_air(&proof_options);
 
-    let air_trace_pairs: Vec<(
-        &dyn AIR<Field = F, FieldExtension = E, PublicInputs = ()>,
-        _,
-        _,
-    )> = vec![
-        (&cpu_air, cpu_trace, &()),
-        (&bitwise_air, bitwise_trace, &()),
-        (&lt_air, lt_trace, &()),
-    ];
+    let air_trace_pairs: Vec<(&dyn AIR<Field = F, FieldExtension = E, PublicInputs = ()>, _, _)> =
+        vec![
+            (&cpu_air, cpu_trace, &()),
+            (&bitwise_air, bitwise_trace, &()),
+            (&lt_air, lt_trace, &()),
+        ];
 
     let multi_proof =
         match Prover::multi_prove(air_trace_pairs, &mut DefaultTranscript::<E>::new(&[])) {
@@ -92,56 +80,6 @@ fn prove_and_verify_vm(
         vec![&cpu_air, &bitwise_air, &lt_air];
 
     Verifier::multi_verify(&airs, &multi_proof, &mut DefaultTranscript::<E>::new(&[]))
-}
-
-/// Run multi_prove and multi_verify for all VM tables with MINIMAL bitwise.
-///
-/// Used for fast tests where the bitwise table is a dummy that only contains
-/// the rows needed to balance the bus. NOT the full preprocessed table.
-fn prove_and_verify_vm_minimal(
-    cpu_trace: &mut TraceTable<F, E>,
-    bitwise_trace: &mut TraceTable<F, E>,
-    lt_trace: &mut TraceTable<F, E>,
-) -> bool {
-    let proof_options = ProofOptions::default_test_options();
-
-    let cpu_air = create_cpu_air(&proof_options);
-    let bitwise_air = create_bitwise_air(&proof_options);
-    let lt_air = create_lt_air(&proof_options);
-
-    let air_trace_pairs: Vec<(
-        &dyn AIR<Field = F, FieldExtension = E, PublicInputs = ()>,
-        _,
-        _,
-    )> = vec![
-        (&cpu_air, cpu_trace, &()),
-        (&bitwise_air, bitwise_trace, &()),
-        (&lt_air, lt_trace, &()),
-    ];
-
-    eprintln!("DEBUG: Proving {} tables...", air_trace_pairs.len());
-    let multi_proof =
-        match Prover::multi_prove(air_trace_pairs, &mut DefaultTranscript::<E>::new(&[])) {
-            Ok(proof) => {
-                eprintln!(
-                    "DEBUG: Prover succeeded, {} proofs generated",
-                    proof.proofs.len()
-                );
-                proof
-            }
-            Err(e) => {
-                eprintln!("Prover error: {:?}", e);
-                return false;
-            }
-        };
-
-    let airs: Vec<&dyn AIR<Field = F, FieldExtension = E, PublicInputs = ()>> =
-        vec![&cpu_air, &bitwise_air, &lt_air];
-
-    eprintln!("DEBUG: Verifying {} AIRs...", airs.len());
-    let result = Verifier::multi_verify(&airs, &multi_proof, &mut DefaultTranscript::<E>::new(&[]));
-    eprintln!("DEBUG: Verification result: {}", result);
-    result
 }
 
 // =============================================================================
